+++ src/main/java/com/sbsolutions/order/models/OrderPlacementStatus.java
@@ -1,0 +1,8 @@
+package com.sbsolutions.order.models;
+
+public enum OrderPlacementStatus {
+  OFF,
+  SCHEDULED,
+  OPEN,
+  CLOSED
+}