+++ src/main/java/com/sbsolutions/order/models/Status.java
@@ -1,7 +1,9 @@
 package com.sbsolutions.order.models;
 
+
 import java.io.Serial;
 import java.io.Serializable;
+import java.time.LocalTime;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
@@ -12,14 +14,16 @@ import org.springframework.data.annotation.Id;
 @RequiredArgsConstructor
 @AllArgsConstructor
 @EqualsAndHashCode(of = {"id"})
-public class Icing implements Serializable {
+public class Status implements Serializable {
 
   @Serial
   private static final long serialVersionUID = 1L;
 
   @Id
-  String id;
-  String description;
-  String url;
-
+  private String id;
+  private OrderPlacementStatus status;
+  private LocalTime startTime;
+  private LocalTime endTime;
+  private String approvedMessage;
+  private String declinedMessage;
 }