+++ src/test/java/com/sbsolutions/order/models/ItemTest.java
@@ -32,4 +32,15 @@ class ItemTest {
     assertThat(item.getItemType()).isEqualTo(ItemType.DONUT);
     assertThat(item.getQty()).isEqualTo(5);
   }
+
+  @Test
+  void testEqualsAndHashCode() {
+    Item item1 = new Item("1", 100, 200, ItemType.DONUT, 5);
+    Item item2 = new Item("1", 101, 201, ItemType.DONUT, 6); // Different data, same ID
+    Item item3 = new Item("2", 100, 200, ItemType.DONUT, 5);
+
+    assertThat(item1).isEqualTo(item2);
+    assertThat(item1).hasSameHashCodeAs(item2);
+    assertThat(item1).isNotEqualTo(item3);
+  }
 }